refactor(travel): isolate the destination catalogue
